Solved problem with z-index and apply button
The 3d transform was interfering with the z-index property, some
logic has been implemented in order to reset the class of the
slider to slider and avoid having the transform property when the
slider is not repositioning itself. This allows us to have only
clickable buttons when the transition has been completed.